# Mykola_Biliaiev/egghead.io/server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Egghead"
  },
  "host": "localhost",
  "basePath": "/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/lessons": {
      "get": {
        "summary": "List all lessons",
        "operationId": "lesson",
        "tags": [
          "lessons"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "An paged array of lessons",
            "headers": {
              "x-next": {
                "type": "string",
                "description": "A link to the next page of responses"
              }
            },
            "schema": {
              "$ref": "#/definitions/lessons"
            }
          },
          "default": {
            "description": "unexpected error"
          }
        }
      },
      "post": {
        "summary": "Create a lesson",
        "operationId": "createLessons",
        "tags": [
          "lessons"
        ],
        "responses": {
          "201": {
            "description": "Null response"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    },
    "/api/courses": {
      "get": {
        "summary": "List all courses",
        "operationId": "courses",
        "tags": [
          "courses"
        ],
        "responses": {
          "200": {
            "description": "An paged array of Course",
            "headers": {
              "x-next": {
                "type": "string",
                "description": "A link to the next page of responses"
              }
            },
            "schema": {
              "$ref": "#/definitions/lessons"
            }
          },
          "default": {
            "description": "unexpected error"
          }
        }
      },
      "post": {
        "summary": "Create a Course",
        "operationId": "createLessons",
        "tags": [
          "courses"
        ],
        "responses": {
          "201": {
            "description": "Null response"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    },
    "/auth/users": {
      "get": {
        "summary": "List all users (Only admin)",
        "operationId": "users",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "An paged array of Users",
            "headers": {
              "x-next": {
                "type": "string",
                "description": "A link to the next page of responses"
              }
            },
            "schema": {
              "$ref": "#/definitions/lessons"
            }
          },
          "default": {
            "description": "unexpected error"
          }
        }
      },
      "post": {
        "summary": "Create a User",
        "operationId": "createLessons",
        "tags": [
          "users"
        ],
        "responses": {
          "201": {
            "description": "Null response"
          },
          "default": {
            "description": "unexpected error"
          }
        }
      }
    }
  },
  "definitions": {
    "lessons": {
      "required": [
        "id",
        "lessonName",
        "author",
        "time",
        "description"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "lessonName": {
          "type": "string"
        },
        "author": {
          "type": "integer"
        },
        "time": {
          "type": "string"
        },
        "video": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "courses": {
      "required": [
        "id",
        "courseName",
        "courseTech",
        "courseDescription",
        "author"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "courseName": {
          "type": "string"
        },
        "courseTech": {
          "type": "string"
        },
        "courseDescription": {
          "type": "string"
        },
        "author": {
          "type": "integer"
        },
        "lessons": {
          "type": "array",
          "description": "an array of Lessons"
        }
      }
    },
    "users": {
      "required": [
        "id",
        "password",
        "name",
        "email"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "icon": {
          "type": "integer"
        }
      }
    }
  }
}